Do all ProjectorSet messages at once
Create a custom socket wrapper to handle connect and shutdown for
the socket. Also wrap close so that we can use the closing context
manager to make life a little simpler.

Add connect method to Projector, which connects and returns a Socket
wrapper in a context manager. Use this in query, which is used by
all of the action methods.

Make ProjectorSet a subclass of this new projector, and override
connect and query so that the methods on Projector will also
work without modification on the ProjectorSet.

The ProjectorSet versions of connect and query create and work on
lists of Socket wrappers, and sends the message to all of the sockets
before waiting for a response from any of the sockets. This allows
the message to get passed immediately, but avoids problems that arise
if you close the connection before waiting for a response.

This allows both projectors to be blanked and unblanked at the same
time, even though the projector takes a second to respond to the
command.