Fix installation errors and add readme
hatch doesn't include html files by default if they aren't in a package,
so we need to manually declare all the files that we want included in
the distribution. Which is easy enough, because we only need two files.

Fix the systemd example service file import path, and add a default
fake driver configuration.

Add a readme that walks through installing this on a Raspberry Pi.